requirements.txt
faiss-cpu
sentence-transformers
python-docx
fastapi
uvicorn
transformers
torch
